Renamed cryptoapi to crypto.api; Moved models to crypto_api module
