+++ bip39/src/main/kotlin/org/kethereum/bip39/Mnemonic.kt
@@ -3,7 +3,7 @@ package org.kethereum.bip39
 import org.kethereum.bip32.model.Seed
 import org.kethereum.bip32.toKey
 import org.kethereum.bip39.model.MnemonicWords
-import org.kethereum.cryptoapi.kdf.pbkdf2
+import org.kethereum.crypto.api.kdf.pbkdf2
 import org.kethereum.extensions.toBitArray
 import org.kethereum.extensions.toByteArray
 import org.kethereum.hashes.sha256
