+++ crypto/src/test/kotlin/org/kethereum/crypto/KeysTest.kt
@@ -1,8 +1,6 @@
 package org.kethereum.crypto
 
 import org.assertj.core.api.Assertions.assertThat
-import org.junit.Assert.assertNotNull
-import org.junit.Before
 import org.junit.Test
 import org.kethereum.crypto.data.ADDRESS
 import org.kethereum.crypto.data.PUBLIC_KEY
@@ -26,7 +24,8 @@ class KeysTest {
 
     @Test
     fun testGetAddressZeroPaddedAddress() {
-        val publicKey = PublicKey("0xa1b31be4d58a7ddd24b135db0da56a90fb5382077ae26b250e1dc9cd6232ce22" + "70f4c995428bc76aa78e522316e95d7834d725efc9ca754d043233af6ca90113")
+        val publicKey =
+            PublicKey("0xa1b31be4d58a7ddd24b135db0da56a90fb5382077ae26b250e1dc9cd6232ce22" + "70f4c995428bc76aa78e522316e95d7834d725efc9ca754d043233af6ca90113")
         assertThat(publicKey.toAddress()).isEqualTo(Address("01c52b08330e05d731e38c856c1043288f7d9744"))
     }
