+++ crypto_api/build.gradle
@@ -1,2 +1,4 @@
 dependencies {
+    implementation project(":extensions")
+    implementation project(":model")
 }
