+++ crypto_api/src/main/kotlin/org/kethereum/crypto/api/ec/CurvePoint.kt
@@ -1,0 +1,14 @@
+package org.kethereum.crypto.api.ec
+
+import java.math.BigInteger
+
+interface CurvePoint {
+    val x: BigInteger
+    val y: BigInteger
+
+    fun mul(n: BigInteger): CurvePoint
+    fun add(p: CurvePoint): CurvePoint
+    fun normalize(): CurvePoint
+    fun isInfinity(): Boolean
+    fun encoded(compressed: Boolean = false): ByteArray
+}
