+++ crypto_api/src/main/kotlin/org/kethereum/crypto/api/ec/ECDSASignature.kt
@@ -1,0 +1,5 @@
+package org.kethereum.crypto.api.ec
+
+import java.math.BigInteger
+
+data class ECDSASignature(val r: BigInteger, val s: BigInteger)
