+++ crypto_api/src/main/kotlin/org/kethereum/crypto/api/ec/KeyPairGenerator.kt
@@ -1,0 +1,9 @@
+package org.kethereum.crypto.api.ec
+
+import java.math.BigInteger
+
+typealias PrivateAndPublicKey = Pair<BigInteger, BigInteger>
+
+interface KeyPairGenerator {
+    fun generate(): PrivateAndPublicKey
+}
