+++ crypto_api/src/main/kotlin/org/kethereum/crypto/api/ec/Signer.kt
@@ -1,0 +1,11 @@
+package org.kethereum.crypto.api.ec
+
+import java.math.BigInteger
+
+interface Signer {
+    fun sign(transactionHash: ByteArray, privateKey: BigInteger): ECDSASignature
+
+    fun recover(recId: Int, sig: ECDSASignature, message: ByteArray?): BigInteger?
+
+    fun publicFromPrivate(privateKey: BigInteger): BigInteger
+}
