+++ crypto_api/src/main/kotlin/org/kethereum/crypto/api/hashing/KeccakDigest256.kt
@@ -1,0 +1,8 @@
+package org.kethereum.crypto.api.hashing
+
+
+interface KeccakDigest256 {
+    fun update(data: ByteArray)
+
+    fun digest(): ByteArray
+}
