+++ crypto_api/src/main/kotlin/org/kethereum/crypto/api/kdf/SCrypt.kt
@@ -1,0 +1,5 @@
+package org.kethereum.crypto.api.kdf
+
+interface SCrypt {
+    fun derive(password: ByteArray, salt: ByteArray?, n: Int, r: Int, p: Int, dklen: Int): ByteArray
+}
