--- crypto_api/src/main/kotlin/org/kethereum/cryptoapi/ec/Curve.kt
@@ -1,11 +1,0 @@
-package org.kethereum.cryptoapi.ec
-
-import java.math.BigInteger
-
-interface Curve {
-    val n: BigInteger
-    val g: CurvePoint
-
-    fun decodePoint(data: ByteArray): CurvePoint
-    fun createPoint(x: BigInteger, y: BigInteger): CurvePoint
-}
