--- crypto_api/src/main/kotlin/org/kethereum/cryptoapi/ec/CurvePoint.kt
@@ -1,14 +1,0 @@
-package org.kethereum.cryptoapi.ec
-
-import java.math.BigInteger
-
-interface CurvePoint {
-    val x: BigInteger
-    val y: BigInteger
-
-    fun mul(n: BigInteger): CurvePoint
-    fun add(p: CurvePoint): CurvePoint
-    fun normalize(): CurvePoint
-    fun isInfinity(): Boolean
-    fun encoded(compressed: Boolean = false): ByteArray
-}
