--- crypto_api/src/main/kotlin/org/kethereum/cryptoapi/ec/ECDSASignature.kt
@@ -1,5 +1,0 @@
-package org.kethereum.cryptoapi.ec
-
-import java.math.BigInteger
-
-data class ECDSASignature(val r: BigInteger, val s: BigInteger)
