--- crypto_api/src/main/kotlin/org/kethereum/cryptoapi/ec/KeyPairGenerator.kt
@@ -1,9 +1,0 @@
-package org.kethereum.cryptoapi.ec
-
-import java.math.BigInteger
-
-typealias PrivateAndPublicKey = Pair<BigInteger, BigInteger>
-
-interface KeyPairGenerator {
-    fun generate(): PrivateAndPublicKey
-}
