--- crypto_api/src/main/kotlin/org/kethereum/cryptoapi/ec/Signer.kt
@@ -1,11 +1,0 @@
-package org.kethereum.cryptoapi.ec
-
-import java.math.BigInteger
-
-interface Signer {
-    fun sign(transactionHash: ByteArray, privateKey: BigInteger): ECDSASignature
-
-    fun recover(recId: Int, sig: ECDSASignature, message: ByteArray?): BigInteger?
-
-    fun publicFromPrivate(privateKey: BigInteger): BigInteger
-}
