--- crypto_api/src/main/kotlin/org/kethereum/cryptoapi/hashing/KeccakDigest256.kt
@@ -1,8 +1,0 @@
-package org.kethereum.cryptoapi.hashing
-
-
-interface KeccakDigest256 {
-    fun update(data: ByteArray)
-
-    fun digest(): ByteArray
-}
