--- crypto_api/src/main/kotlin/org/kethereum/cryptoapi/kdf/SCrypt.kt
@@ -1,5 +1,0 @@
-package org.kethereum.cryptoapi.kdf
-
-interface SCrypt {
-    fun derive(password: ByteArray, salt: ByteArray?, n: Int, r: Int, p: Int, dklen: Int): ByteArray
-}
