+++ crypto_api_wrapper/src/main/kotlin/org/kethereum/crypto/api/ec/ECWrappers.kt
@@ -1,0 +1,7 @@
+package org.kethereum.crypto.api.ec
+
+fun curve(): Curve = EllipticCurve
+
+fun signer(): Signer = EllipticCurveSigner
+
+fun keyPairGenerator(): KeyPairGenerator = EllipticCurveKeyPairGenerator
