+++ crypto_api_wrapper/src/main/kotlin/org/kethereum/crypto/api/hashing/HashingWrappers.kt
@@ -1,0 +1,3 @@
+package org.kethereum.crypto.api.hashing
+
+fun keccackDigest256(): KeccakDigest256 = KeccakDigest256Impl()
