+++ crypto_api_wrapper/src/main/kotlin/org/kethereum/crypto/api/kdf/KDFWrapper.kt
@@ -1,0 +1,5 @@
+package org.kethereum.crypto.api.kdf
+
+fun pbkdf2(): PBKDF2 = PBKDF2Impl
+
+fun scrypt(): SCrypt = SCryptImpl
