--- crypto_api_wrapper/src/main/kotlin/org/kethereum/cryptoapi/ec/ECWrappers.kt
@@ -1,7 +1,0 @@
-package org.kethereum.cryptoapi.ec
-
-fun curve(): Curve = EllipticCurve
-
-fun signer(): Signer = EllipticCurveSigner
-
-fun keyPairGenerator(): KeyPairGenerator = EllipticCurveKeyPairGenerator
