--- crypto_api_wrapper/src/main/kotlin/org/kethereum/cryptoapi/hashing/HashingWrappers.kt
@@ -1,3 +1,0 @@
-package org.kethereum.cryptoapi.hashing
-
-fun keccackDigest256(): KeccakDigest256 = KeccakDigest256Impl()
