--- crypto_api_wrapper/src/main/kotlin/org/kethereum/cryptoapi/kdf/KDFWrapper.kt
@@ -1,5 +1,0 @@
-package org.kethereum.cryptoapi.kdf
-
-fun pbkdf2(): PBKDF2 = PBKDF2Impl
-
-fun scrypt(): SCrypt = SCryptImpl
