+++ eip155/build.gradle
@@ -1,6 +1,7 @@
 dependencies {
     implementation project(":model")
     implementation project(":crypto")
+    implementation project(":crypto_api")
     implementation project(":functions")
     implementation project(":rlp")
