+++ keccak_shortcut/src/main/kotlin/org/kethereum/keccakshortcut/Keccak.kt
@@ -1,6 +1,6 @@
 package org.kethereum.keccakshortcut
 
-import org.kethereum.cryptoapi.hashing.keccackDigest256
+import org.kethereum.crypto.api.hashing.keccackDigest256
 import org.walleth.khex.hexToByteArray
 
 fun String.keccak() = hexToByteArray().keccak()
